ai-hello-world/requirements.txt
fastapi
uvicorn
python-dotenv
langchain-openai
langchain-community
langchain-core
pydantic
requests
dashscope